+++ app-backend/src/__tests__/utils.test.ts
@@ -1,49 +1,58 @@
+/*
+* Copyright (c) 2011-Present, Precise.ly, Inc.
+* All rights reserved.
+*
+* Redistribution and use in source and binary forms, with or
+* without modification, are not permitted.
+*/
+
 jest.mock('aws-sdk', () => {
     return {
         KMS: () => ({
-            decrypt: jest.fn((
-                        params: AWS.KMS.Types.DecryptRequest, 
+            decrypt: jest.fn()
+                .mockImplementationOnce((
+                        params: AWS.KMS.Types.DecryptRequest,
+                        callback: (error: Error, data: AWS.KMS.Types.DecryptResponse) => void
+                ): void => {
+                    callback(new Error('mock error'), null);
+                })
+                .mockImplementationOnce((
+                        params: AWS.KMS.Types.DecryptRequest,
                         callback: (error: Error, data: AWS.KMS.Types.DecryptResponse) => void
                 ): void => {
                     callback(null, {Plaintext: new Buffer(JSON.stringify({demo: 'test'}))});
                 })
-                .mockImplementationOnce((
-                    params: AWS.KMS.Types.DecryptRequest, 
-                    callback: (error: Error, data: AWS.KMS.Types.DecryptResponse) => void
-                ): void => {
-                    callback(new Error('mock error'), null)
-                })
         })
-    }
+    };
 });
 
-import {decryptKMS} from '../utils';
+import {getEnvironmentVariables} from '../utils';
 
 const unroll = require('unroll');
 unroll.use(it);
 
-describe('Test for decryptKMS', (): void => {
-    it('decryptKMS should be a function', (): void => {
-        expect(typeof decryptKMS).toBe('function');
+describe('Test for getEnvironmentVariables', (): void => {
+
+    process.env.SECRETS = 'test';
+
+    it('getEnvironmentVariables should be a function', (): void => {
+        expect(typeof getEnvironmentVariables).toBe('function');
     });
 
-    unroll('it should #expectedResult when #description', (
+    unroll('it should #description when #expectedResult', async (
             done: () => void,
             args: {expectedResult: string, description: string}
-    ): void => {
-        decryptKMS('testData')
-            .then((data: {demo: string}) => {
-                expect(data).toBe({demo: 'test'});
-            })
-            .catch((error: Error) => {
-                expect(error.message).toBe('mock error');
-            })
-            .finally(() => {
-                done();
-            });
+    ): Promise<void> => {
+        const result = await getEnvironmentVariables();
+        if(args.expectedResult === 'pass') {
+            expect(result).toEqual({"demo": "test"});
+        } else {
+            expect(result).toBeFalsy();
+        }
+        done();
     }, [
         ['expectedResult', 'description'],
-        ['fail', 'KMS.decrypt throws error'],
-        ['pass', 'KMS.decrypt calls success callback'],
-    ])
+        ['fail', 'return null'],
+        ['pass', 'return data'],
+    ]);
 });