+++ app-client/src/containers/user/Dashboard.tsx
@@ -9,8 +9,8 @@
 import * as React from 'react';
 import * as Radium from 'radium';
 import {RouteComponentProps} from 'react-router';
-import {Button} from 'src/components/ReusableComponents';
-import {logOut} from 'src/utils/cognito';
+import {Container} from 'src/components/ReusableComponents';
+import {NavigationBar} from 'src/components/navigationBar/NavigationBar';
 
 @Radium
 export class Dashboard extends React.Component<RouteComponentProps<void>> {
@@ -23,10 +23,11 @@ export class Dashboard extends React.Component<RouteComponentProps<void>> {
     render(): JSX.Element {
         return (
             <div>
-                <h6>This is a dummy dashboard page.</h6><br/>
-                <Button onClick={(logout): void => { logOut(); this.props.history.push('/login'); }}>
-                    Logout
-                </Button>
+                <NavigationBar {...this.props}/>
+                <hr/>
+                <Container>
+                    <h6>This is a dummy dashboard page.</h6><br/>
+                </Container>
             </div>
         );
     }